Move ResourceContainer to plugin package
